splitted inventory movement from the character and added merchant move inventory proto response
